Feat: check and rename input data.frames
Adds the check and rename function to utils.
Calls the check and rename function in all the shift functions.
Adds a test file to check the check and rename function.
Updates some code that is no longer needed becuase of the check and rename function.
Updates a bit of shifterator documentation to reflect the terminology of R.