FEAT: Syncronised y_limits and collapse x_lables on the total plot
The plotting parameters for detailed or not will now calculate the y_limits for the total contribution plot and the (detailed) main plot. Calculation is now done in one place.

Created x_labels for the total contribution plot so that bars are plotted on the same line where the values are the opposites of each other.